requirements.txt
# requirements.txt

# Core Streamlit Framework & UI Components
streamlit
streamlit-option-menu
streamlit_js_eval

# Authentication & Security
streamlit-authenticator>=0.3.0
passlib
bcrypt
werkzeug

# Database & ORM
sqlalchemy
psycopg2-binary
alembic

# Data Handling & Utilities
pandas
numpy
PyYAML
python-dotenv
openpyxl
